+++ app/pages/docs/DocsPage.tsx
@@ -29,10 +29,12 @@ import Tkc from './Tkc';
 import Practice from './Practice';
 import DocsContainer from './DocsContainer';
 import FiveQ from './5qcard/FiveQ';
+import FourS from './FourS';
 
 export enum Pages {
     DefaultPage,
     FiveQCard,
+    FourSteps,
     KtsPage,
     PracticePage,
     TkcPage,
@@ -54,6 +56,7 @@ const DocsPage: React.FunctionComponent = () => {
                         }}
                     >
                         {status === Pages.FiveQCard && <FiveQ />}
+                        {status === Pages.FourSteps && <FourS />}
                         {status === Pages.TkcPage && <Tkc />}
                         {status === Pages.PracticePage && <Practice />}
                         {status === Pages.KtsPage && <Kts />}